+++ CHANGES.md
@@ -1,5 +1,11 @@
 # CHANGES
 
+## Not released:
+
+- fix rename additional checks to include [#7](https://github.com/mrijken/check-config/issues/7)
+- fix relative includes [#8](https://github.com/mrijken/check-config/issues/8)
+
+
 ## 0.6.1
 
  - BREAKING: Use check_config.toml (in stead of checkers.toml) as default style file